r2900/sofa-dev : FIX : warnings

+++ modules/sofa/component/topology/Hexa2QuadTopologicalMapping.cpp
@@ -237,7 +237,7 @@ void Hexa2QuadTopologicalMapping::updateTopologicalMapping()
 
                                 ind_real_last = Glob2LocMap[last];
 
-                                if(k != last)
+                                if((int) k != last)
                                 {
 
                                     Glob2LocMap.erase(Glob2LocMap.find(k));
@@ -252,7 +252,7 @@ void Hexa2QuadTopologicalMapping::updateTopologicalMapping()
                                 }
                             }
 
-                            if( ind_k != ind_last)
+                            if((int) ind_k != ind_last)
                             {
 
                                 Glob2LocMap.erase(Glob2LocMap.find(Loc2GlobVec[ind_last]));
